Move UnitsHelper out of core/ to correct ODR violation
Previously, under src/core/mediator, UnitsHelper.cpp was being
compiled and linked twice, in CMake targets NGen::core and
NGen::core_mediator. This occurred because their CMakeLists.txt use
dynamic_sourced_cxx_library, which is implemented in terms of
GLOB_RECURSE. Having two copies of the object file in two separate
library archives meant that different objects could end up resolving
symbols from either one, depending on link order.

That's a violation of the C++ One Definition Rule (ODR) no matter
what. In this case, it had the easy potential to directly fail because of the
static variables contained in that compilation unit.
