+++ CMakeLists.txt
@@ -332,6 +332,7 @@ endif()
 target_include_directories(ngen PUBLIC "${NGEN_INC_DIR}")
 
 add_subdirectory("src/core")
+add_subdirectory("src/mediator")
 add_subdirectory("src/geojson")
 add_subdirectory("src/bmi")
 add_subdirectory("src/realizations/catchment")
