+++ include/realizations/catchment/Bmi_Multi_Formulation.hpp
@@ -11,7 +11,7 @@
 #include "ConfigurationException.hpp"
 #include "ExternalIntegrationException.hpp"
 #include "utilities/logging_utils.h"
-#include <core/mediator/UnitsHelper.hpp>
+#include <mediator/UnitsHelper.hpp>
 
 #define BMI_REALIZATION_CFG_PARAM_REQ__MODULES "modules"
 #define BMI_REALIZATION_CFG_PARAM_OPT__DEFAULT_OUT_VALS "default_output_values"
