+++ src/mediator/CMakeLists.txt
@@ -11,8 +11,7 @@ find_package(Threads REQUIRED)
 
 target_include_directories(core_mediator PUBLIC
         ${PROJECT_SOURCE_DIR}/include
-        ${PROJECT_SOURCE_DIR}/include/core
-        ${PROJECT_SOURCE_DIR}/include/core/mediator
+        ${PROJECT_SOURCE_DIR}/include/mediator
         )
 
 target_link_libraries(core_mediator PUBLIC
